refactor: home the longitudinal grid in propagation
_propagation_grid/_valid_window/_node_grid/_node_geometry cluster with
propagation.py's functions in the call graph -- they are path geometry and
quadrature, not Cherenkov orchestration. Move them there as pure functions
(propagation_grid, visibility_window, gl_node_grid, node_geometry). The
three shower-physics depths they consumed (e2hill age-floor, Greisen death
depth, shower-maximum depth) are now explicit parameters computed by a thin
CphotAng._propagation_grid wrapper, so the propagation layer stays free of
shower models and no import cycle forms.

Also homes the generic cached_leggauss GL cache in quadrature.py, replacing
four deferred function-local imports propagation used to dodge the old
layering. Goldens bit-identical.
